Refactor filter components by moving styled elements to FilterContainers.tsx and updating imports in ChainsContent.tsx and PlatformsContent.tsx. This improves code organization and readability.
